Backend/requirements.txt
# Flask and Core Extensions
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.0

# Database
psycopg2-binary==2.9.10  # PostgreSQL database driver 

# Data Processing and Analysis
pandas
numpy

# Web Scraping and API Requests
gnews==0.3.9
requests==2.32.3

# Sentiment Analysis and NLP
nltk==3.9.1
transformers==4.47.0  # For advanced NLP (e.g., OpenAI GPT models)
torch==2.5.1  # PyTorch for transformer-based models (if needed)

# Utility Libraries
python-dotenv==1.0.1  # For environment variable management
validators==0.34.0  # For input validation

# Testing
pytest==8.3.4
pytest-flask==1.3.0
coverage==7.6.9